splits a cli test into multiple files
