+++ aqueduct/test/command/migration_test.dart
@@ -10,10 +10,8 @@ import 'package:aqueduct/src/cli/mixins/project.dart';
 import 'package:test/test.dart';
 
 import '../helpers.dart';
-import 'cli_helpers.dart';
 
 void main() {
-  justLogEverything();
   group("Cooperation", () {
     PersistentStore store;
 
@@ -136,96 +134,6 @@ class Migration1 extends Migration { @override Future upgrade() async {} @overri
       expect(files[2].uri.pathSegments.last, "03_Foo.migration.dart");
       expect(files[3].uri.pathSegments.last, "000001001.migration.dart");
       expect(files[4].uri.pathSegments.last, "10001_.migration.dart");
-    });
-  });
-
-  group("Validating", () {
-    Terminal terminal;
-
-    setUp(() async {
-      terminal = await Terminal.createProject();
-      terminal.addOrReplaceFile("lib/application_test.dart", """
-class TestObject extends ManagedObject<_TestObject> {}
-
-class _TestObject {
-  @primaryKey
-  int id;
-
-  String foo;
-}
-      """);
-      await terminal.getDependencies(offline: true);
-    });
-
-    tearDown(Terminal.deleteTemporaryDirectory);
-
-    test("If validating with no migration dir, get error", () async {
-      var res = await terminal.runAqueductCommand("db", ["validate"]);
-
-      expect(res, isNot(0));
-      expect(terminal.output, contains("No migration files found"));
-    });
-
-    test("Validating two equal schemas succeeds", () async {
-      var res = await terminal.runAqueductCommand("db", ["generate"]);
-      expect(res, 0);
-
-      res = await terminal.runAqueductCommand("db", ["validate"]);
-      expect(res, 0);
-      expect(terminal.output, contains("Validation OK"));
-      expect(terminal.output, contains("version is 1"));
-    });
-
-    test("Validating different schemas fails", () async {
-      var res = await terminal.runAqueductCommand("db", ["generate"]);
-      expect(res, 0);
-
-      terminal.modifyFile("migrations/00000001_initial.migration.dart",
-          (contents) {
-        const upgradeLocation = "upgrade()";
-        final nextLine =
-            contents.indexOf("\n", contents.indexOf(upgradeLocation));
-        return contents.replaceRange(nextLine, nextLine + 1, """
-        database.createTable(SchemaTable(\"foo\", []));
-        """);
-      });
-
-      res = await terminal.runAqueductCommand("db", ["validate"]);
-      expect(res, isNot(0));
-      expect(terminal.output, contains("Validation failed"));
-    });
-
-    test(
-        "Validating runs all migrations in directory and checks the total product",
-        () async {
-      var res = await terminal.runAqueductCommand("db", ["generate"]);
-      expect(res, 0);
-
-      terminal.modifyFile("migrations/00000001_initial.migration.dart",
-          (contents) {
-        const upgradeLocation = "upgrade()";
-        final nextLine =
-            contents.indexOf("\n", contents.indexOf(upgradeLocation));
-        return contents.replaceRange(nextLine, nextLine + 1, """
-        database.createTable(SchemaTable(\"foo\", []));
-        """);
-      });
-
-      res = await terminal.runAqueductCommand("db", ["validate"]);
-      expect(res, isNot(0));
-      expect(terminal.output, contains("Validation failed"));
-
-      res = await terminal.runAqueductCommand("db", ["generate"]);
-      expect(res, 0);
-
-      var secondMigrationFile = File.fromUri(terminal
-          .defaultMigrationDirectory.uri
-          .resolve("00000002_unnamed.migration.dart"));
-      expect(secondMigrationFile.readAsStringSync(),
-          contains("database.deleteTable(\"foo\")"));
-
-      res = await terminal.runAqueductCommand("db", ["validate"]);
-      expect(res, 0);
     });
   });
 }
